+++ src/test/scala/edu/ie3/simona/service/em/ExtEmDataServiceSpec.scala
@@ -14,25 +14,18 @@ import edu.ie3.simona.api.data.model.em.{EmSetPoint, FlexOptionRequest}
 import edu.ie3.simona.api.ontology.ScheduleDataServiceMessage
 import edu.ie3.simona.api.ontology.em.*
 import edu.ie3.simona.api.ontology.simulation.ControlResponseMessageFromExt
-import edu.ie3.simona.ontology.messages.SchedulerMessage.{
-  Completion,
-  ScheduleActivation,
-}
+import edu.ie3.simona.ontology.messages.SchedulerMessage.Completion
 import edu.ie3.simona.ontology.messages.ServiceMessage.{
-  Create,
   EmFlexMessage,
   EmServiceRegistration,
 }
 import edu.ie3.simona.ontology.messages.flex.FlexibilityMessage.*
 import edu.ie3.simona.ontology.messages.flex.{FlexType, PowerLimitFlexOptions}
 import edu.ie3.simona.ontology.messages.{Activation, SchedulerMessage}
-import edu.ie3.simona.scheduler.ScheduleLock
 import edu.ie3.simona.service.DataTimeType
 import edu.ie3.simona.service.em.ExtEmDataService.InitExtEmData
-import edu.ie3.simona.test.common.TestSpawnerTyped
 import edu.ie3.simona.test.common.input.EmInputTestData
 import edu.ie3.simona.util.CollectionUtils.asJava
-import edu.ie3.simona.util.SimonaConstants.INIT_SIM_TICK
 import edu.ie3.util.quantities.QuantityUtils.{asKiloWatt, asMegaWatt}
 import edu.ie3.util.scala.quantities.DefaultQuantities.zeroKW
 import org.apache.pekko.actor.testkit.typed.scaladsl.{
@@ -51,8 +44,7 @@ import scala.jdk.CollectionConverters.*
 class ExtEmDataServiceSpec
     extends ScalaTestWithActorTestKit
     with AnyWordSpecLike
-    with EmInputTestData
-    with TestSpawnerTyped {
+    with EmInputTestData {
 
   implicit val simulationStart: ZonedDateTime = ZonedDateTime.now()
 
@@ -69,31 +61,21 @@ class ExtEmDataServiceSpec
       val extSimAdapter =
         TestProbe[ControlResponseMessageFromExt]("extSimAdapter")
 
-      val emService = spawn(ExtEmDataService(scheduler.ref))
       val extEmDataConnection =
         new ExtEmDataConnection(emptyControlled, EmMode.BASE)
+      val emService = spawn(
+        ExtEmDataService(
+          scheduler.ref,
+          InitExtEmData(extEmDataConnection, simulationStart),
+        )
+      )
 
       extEmDataConnection.setActorRefs(
         emService,
         extSimAdapter.ref,
       )
 
-      val key =
-        ScheduleLock.singleKey(TSpawner, scheduler.ref, INIT_SIM_TICK)
-      scheduler
-        .expectMessageType[ScheduleActivation] // lock activation scheduled
-
-      emService ! Create(
-        InitExtEmData(extEmDataConnection, simulationStart),
-        key,
-      )
-
-      scheduler.expectMessage(
-        ScheduleActivation(emService, INIT_SIM_TICK, Some(key))
-      )
-
-      emService ! Activation(INIT_SIM_TICK)
-      scheduler.expectMessage(Completion(emService))
+      scheduler.expectNoMessage()
     }
 
     "stash registration request and handle it correctly once initialized" in {
@@ -101,9 +83,14 @@ class ExtEmDataServiceSpec
       val extSimAdapter =
         TestProbe[ControlResponseMessageFromExt]("extSimAdapter")
 
-      val emService = spawn(ExtEmDataService(scheduler.ref))
       val extEmDataConnection =
         new ExtEmDataConnection(emptyControlled, EmMode.BASE)
+      val emService = spawn(
+        ExtEmDataService(
+          scheduler.ref,
+          InitExtEmData(extEmDataConnection, simulationStart),
+        )
+      )
 
       extEmDataConnection.setActorRefs(
         emService,
@@ -121,23 +108,6 @@ class ExtEmDataServiceSpec
       )
 
       scheduler.expectNoMessage()
-
-      val key =
-        ScheduleLock.singleKey(TSpawner, scheduler.ref, INIT_SIM_TICK)
-      scheduler
-        .expectMessageType[ScheduleActivation] // lock activation scheduled
-
-      emService ! Create(
-        InitExtEmData(extEmDataConnection, simulationStart),
-        key,
-      )
-
-      scheduler.expectMessage(
-        ScheduleActivation(emService, INIT_SIM_TICK, Some(key))
-      )
-
-      emService ! Activation(INIT_SIM_TICK)
-      scheduler.expectMessage(Completion(emService))
     }
   }
 
@@ -150,30 +120,21 @@ class ExtEmDataServiceSpec
       val extSimAdapter =
         TestProbe[ControlResponseMessageFromExt]("extSimAdapter")
 
-      val emService = spawn(ExtEmDataService(scheduler.ref))
       val extEmDataConnection =
         new ExtEmDataConnection(emptyControlled, EmMode.BASE)
+      val emService = spawn(
+        ExtEmDataService(
+          scheduler.ref,
+          InitExtEmData(extEmDataConnection, simulationStart),
+        )
+      )
 
       extEmDataConnection.setActorRefs(
         emService,
         extSimAdapter.ref,
       )
 
-      val key =
-        ScheduleLock.singleKey(TSpawner, scheduler.ref, INIT_SIM_TICK)
-      scheduler
-        .expectMessageType[ScheduleActivation] // lock activation scheduled
-
-      emService ! Create(
-        InitExtEmData(extEmDataConnection, simulationStart),
-        key,
-      )
-
-      scheduler.expectMessage(
-        ScheduleActivation(emService, INIT_SIM_TICK, Some(key))
-      )
-
-      emService ! Activation(INIT_SIM_TICK)
+      scheduler.expectNoMessage()
 
       emService ! EmServiceRegistration(
         emAgent.ref,
@@ -186,8 +147,6 @@ class ExtEmDataServiceSpec
         FlexCompletion(emAgent1UUID, requestAtTick = Some(0)),
         emAgent1UUID,
       )
-
-      scheduler.expectMessage(Completion(emService))
 
       // we trigger em service and expect an exception
       emService ! Activation(0)
@@ -202,31 +161,21 @@ class ExtEmDataServiceSpec
       val extSimAdapter =
         TestProbe[ControlResponseMessageFromExt]("extSimAdapter")
 
-      val emService = spawn(ExtEmDataService(scheduler.ref))
       val extEmDataConnection =
         new ExtEmDataConnection(emptyControlled, EmMode.BASE)
+      val emService = spawn(
+        ExtEmDataService(
+          scheduler.ref,
+          InitExtEmData(extEmDataConnection, simulationStart),
+        )
+      )
 
       extEmDataConnection.setActorRefs(
         emService,
         extSimAdapter.ref,
       )
 
-      val key =
-        ScheduleLock.singleKey(TSpawner, scheduler.ref, INIT_SIM_TICK)
-      scheduler
-        .expectMessageType[ScheduleActivation] // lock activation scheduled
-
-      emService ! Create(
-        InitExtEmData(extEmDataConnection, simulationStart),
-        key,
-      )
-
-      scheduler.expectMessage(
-        ScheduleActivation(emService, INIT_SIM_TICK, Some(key))
-      )
-
-      emService ! Activation(INIT_SIM_TICK)
-      scheduler.expectMessage(Completion(emService))
+      scheduler.expectNoMessage()
 
       val emAgent1 = TestProbe[EmAgent.Message]("emAgent1")
       val emAgent2 = TestProbe[EmAgent.Message]("emAgent2")
@@ -322,31 +271,21 @@ class ExtEmDataServiceSpec
       val extSimAdapter =
         TestProbe[ControlResponseMessageFromExt]("extSimAdapter")
 
-      val emService = spawn(ExtEmDataService(scheduler.ref))
       val extEmDataConnection =
         new ExtEmDataConnection(emptyControlled, EmMode.BASE)
+      val emService = spawn(
+        ExtEmDataService(
+          scheduler.ref,
+          InitExtEmData(extEmDataConnection, simulationStart),
+        )
+      )
 
       extEmDataConnection.setActorRefs(
         emService,
         extSimAdapter.ref,
       )
 
-      val key =
-        ScheduleLock.singleKey(TSpawner, scheduler.ref, INIT_SIM_TICK)
-      scheduler
-        .expectMessageType[ScheduleActivation] // lock activation scheduled
-
-      emService ! Create(
-        InitExtEmData(extEmDataConnection, simulationStart),
-        key,
-      )
-
-      scheduler.expectMessage(
-        ScheduleActivation(emService, INIT_SIM_TICK, Some(key))
-      )
-
-      emService ! Activation(INIT_SIM_TICK)
-      scheduler.expectMessage(Completion(emService))
+      scheduler.expectNoMessage()
 
       val emAgent1 = TestProbe[EmAgent.Message]("emAgent1")
       val emAgent2 = TestProbe[EmAgent.Message]("emAgent2")