+++ src/test/scala/edu/ie3/simona/service/load/LoadProfileServiceSpec.scala
@@ -17,12 +17,10 @@ import edu.ie3.simona.ontology.messages.SchedulerMessage.{
 }
 import edu.ie3.simona.ontology.messages.ServiceMessage.*
 import edu.ie3.simona.ontology.messages.{Activation, SchedulerMessage}
-import edu.ie3.simona.scheduler.ScheduleLock
 import edu.ie3.simona.service.Data.SecondaryData.{LoadData, SecondarySeriesData}
 import edu.ie3.simona.service.DataTimeType
 import edu.ie3.simona.service.load.LoadProfileService.InitLoadProfileServiceStateData
 import edu.ie3.simona.test.common.{ConfigTestData, TestSpawnerTyped}
-import edu.ie3.simona.util.SimonaConstants.INIT_SIM_TICK
 import edu.ie3.util.TimeUtil
 import org.apache.pekko.actor.testkit.typed.scaladsl.{
   ScalaTestWithActorTestKit,
@@ -56,30 +54,20 @@ class LoadProfileServiceSpec
 
   // build the load profile service
   private val loadProfileService = testKit.spawn(
-    LoadProfileService(scheduler.ref)
+    LoadProfileService(
+      scheduler.ref,
+      InitLoadProfileServiceStateData(
+        sourceDefinition,
+        TimeUtil.withDefaults.toZonedDateTime(
+          simonaConfig.time.startDateTime
+        ),
+      ),
+    )
   )
 
   "A load profile service" should {
-    "send correct completion message after initialisation" in {
-      val key =
-        ScheduleLock.singleKey(TSpawner, scheduler.ref, INIT_SIM_TICK)
-      scheduler
-        .expectMessageType[ScheduleActivation] // lock activation scheduled
-
-      loadProfileService ! Create(
-        InitLoadProfileServiceStateData(
-          sourceDefinition,
-          simonaConfig.time.simStartTime,
-        ),
-        key,
-      )
-
-      val activationMsg = scheduler.expectMessageType[ScheduleActivation]
-      activationMsg.tick shouldBe INIT_SIM_TICK
-      activationMsg.unlockKey shouldBe Some(key)
-
-      loadProfileService ! Activation(INIT_SIM_TICK)
-      scheduler.expectMessage(Completion(activationMsg.actor, Some(0)))
+    "send correct schedule message after initialisation" in {
+      scheduler.expectMessage(ScheduleActivation(loadProfileService, 0L))
     }
 
     "announce failed load profile registration on invalid load profile" in {